requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
gunicorn
dj-database-url
django-cors-headers
